Update links to fee information doc pages (#588)
* update fee info url

* add fee info link to amountless btc address page

remove estimated fee message

* extract amountless BTC Address page widgets
